Move remote tensor from backend/device to utils
