+++ src/plugins/intel_npu/src/backend/include/zero_device.hpp
@@ -38,22 +38,6 @@ public:
         log.setLevel(config.get<LOG_LEVEL>());
     }
 
-    ov::SoPtr<ov::IRemoteTensor> createRemoteTensor(
-        std::shared_ptr<ov::IRemoteContext> context,
-        const ov::element::Type& element_type,
-        const ov::Shape& shape,
-        const Config& config,
-        ov::intel_npu::TensorType tensor_type = ov::intel_npu::TensorType::BINDED,
-        ov::intel_npu::MemType mem_type = ov::intel_npu::MemType::L0_INTERNAL_BUF,
-        void* mem = nullptr) override;
-
-    ov::SoPtr<ov::ITensor> createHostTensor(
-        std::shared_ptr<ov::IRemoteContext> context,
-        const ov::element::Type& element_type,
-        const ov::Shape& shape,
-        const Config& config,
-        ov::intel_npu::TensorType tensor_type = ov::intel_npu::TensorType::BINDED) override;
-
     ZeroDevice& operator=(const ZeroDevice&) = delete;
     ZeroDevice(const ZeroDevice&) = delete;
