+++ src/plugins/intel_npu/src/backend/include/zero_infer_request.hpp
@@ -10,10 +10,10 @@
 #include "intel_npu/common/npu.hpp"
 #include "intel_npu/common/sync_infer_request.hpp"
 #include "intel_npu/utils/logger/logger.hpp"
+#include "intel_npu/utils/zero/zero_remote_tensor.hpp"
 #include "intel_npu/utils/zero/zero_utils.hpp"
 #include "intel_npu/utils/zero/zero_wrappers.hpp"
 #include "zero_pipeline.hpp"
-#include "zero_remote_tensor.hpp"
 #include "zero_tensor.hpp"
 
 namespace intel_npu {
