+++ src/plugins/intel_npu/src/backend/src/zero_device.cpp
@@ -7,9 +7,7 @@
 #include "intel_npu/common/itt.hpp"
 #include "intel_npu/utils/zero/zero_api.hpp"
 #include "intel_npu/utils/zero/zero_utils.hpp"
-#include "zero_host_tensor.hpp"
 #include "zero_infer_request.hpp"
-#include "zero_remote_tensor.hpp"
 
 using namespace intel_npu;
 
@@ -176,35 +174,3 @@ std::shared_ptr<SyncInferRequest> ZeroDevice::createInferRequest(
     const Config& config) {
     return std::make_shared<ZeroInferRequest>(_initStructs, compiledModel, config);
 }
-
-ov::SoPtr<ov::IRemoteTensor> ZeroDevice::createRemoteTensor(std::shared_ptr<ov::IRemoteContext> context,
-                                                            const ov::element::Type& element_type,
-                                                            const ov::Shape& shape,
-                                                            const Config& config,
-                                                            ov::intel_npu::TensorType tensor_type,
-                                                            ov::intel_npu::MemType mem_type,
-                                                            void* mem) {
-    return {std::make_shared<ZeroRemoteTensor>(context,
-                                               _initStructs,
-                                               device_properties,
-                                               element_type,
-                                               shape,
-                                               config,
-                                               tensor_type,
-                                               mem_type,
-                                               mem)};
-};
-
-ov::SoPtr<ov::ITensor> ZeroDevice::createHostTensor(std::shared_ptr<ov::IRemoteContext> context,
-                                                    const ov::element::Type& element_type,
-                                                    const ov::Shape& shape,
-                                                    const Config& config,
-                                                    ov::intel_npu::TensorType tensor_type) {
-    return {std::make_shared<ZeroHostTensor>(context,
-                                             _initStructs,
-                                             device_properties,
-                                             element_type,
-                                             shape,
-                                             config,
-                                             tensor_type)};
-};
