+++ src/plugins/intel_npu/src/backend/src/zero_variable_state.cpp
@@ -5,8 +5,8 @@
 #include "zero_variable_state.hpp"
 
 #include "intel_npu/config/options.hpp"
+#include "intel_npu/utils/zero/zero_remote_tensor.hpp"
 #include "intel_npu/utils/zero/zero_utils.hpp"
-#include "zero_remote_tensor.hpp"
 
 namespace intel_npu {
