+++ src/plugins/intel_npu/src/utils/include/intel_npu/utils/zero/zero_host_tensor.hpp
@@ -4,7 +4,6 @@
 
 #pragma once
 
-#include "intel_npu/config/config.hpp"
 #include "intel_npu/utils/zero/zero_init.hpp"
 #include "openvino/runtime/itensor.hpp"
 #include "zero_remote_tensor.hpp"
@@ -15,10 +14,8 @@ class ZeroHostTensor : public ov::ITensor {
 public:
     ZeroHostTensor(const std::shared_ptr<ov::IRemoteContext>& context,
                    const std::shared_ptr<ZeroInitStructsHolder>& init_structs,
-                   const ze_device_properties_t& device_properties,
                    const ov::element::Type element_type,
                    const ov::Shape& shape,
-                   const Config& config,
                    ov::intel_npu::TensorType tensor_type = ov::intel_npu::TensorType::BINDED);
 
     ~ZeroHostTensor() override = default;
